added to ps3linux-install script

+++ PS3LINUX_live_iso_gen.sh
@@ -2,7 +2,6 @@
 
 set -eo pipefail
 
-# Check if root
 if (( EUID != 0 )); then
     echo "Error: This script must be run as root." >&2
     exit 1
@@ -13,16 +12,29 @@ KERNEL_BUILD_PATH="$(pwd)/FC28-x86_64_chroot"
 LIVE_ISO_PATH="$(pwd)/PS3LINUX_LIVE_ISO"
 INITRAMFS_PATH="$(pwd)/initramfs"
 RESOURCES_PATH="$(pwd)/resources"
-EXCLUDES="NetworkManager,NetworkManager-libnm,libndp,sssd-common,sssd-client,libtevent,c-ares,http-parser,jansson,libdhash,libldb,libsss_certmap,libsss_idmap,libsss_nss_idmap,libtalloc,libtdb,ppc64-utils,kernel-bootwrapper,libservicelog,lsvpd,perl-Data-Dumper,perl-Errno,perl-Exporter,perl-File-Path,perl-IO,perl-PathTools,perl-Scalar-List-Utils,perl-Socket,perl-Text-Tabs+Wrap,perl-Unicode-Normalize,perl-constant,perl-interpreter,perl-libs,perl-macros,perl-parent,perl-threads,perl-threads-shared,powerpc-utils,powerpc-utils-core,bc,binutils,librtas,libvpd,perl-Carp,sg3_utils-libs,passwd,libuser"
+EXCLUDES="NetworkManager,NetworkManager-libnm,libndp,sssd-common,sssd-client,libtevent,c-ares,http-parser,jansson,libdhash,libldb,libsss_certmap,libsss_idmap,libsss_nss_idmap,libtalloc,libtdb,ppc64-utils,kernel-bootwrapper,libservicelog,lsvpd,perl-Data-Dumper,perl-Errno,perl-Exporter,perl-File-Path,perl-IO,perl-PathTools,perl-Scalar-List-Utils,perl-Socket,perl-Text-Tabs+Wrap,perl-Unicode-Normalize,perl-constant,perl-interpreter,perl-libs,perl-macros,perl-parent,perl-threads,perl-threads-shared,powerpc-utils,powerpc-utils-core,bc,binutils,librtas,libvpd,perl-Carp,sg3_utils-libs,passwd,libuser,dracut,dracut-config-rescue,plymouth-scripts,firewalld,python3-dbus.python3-firewall,python3-slip,python3-slip-dbus,dbus-glib,ebtables,firewalld-filesystem,ipset,ipset-libs,iptables,libnetfilter_conntrack,libnfnetlink,python3-decorator,python3-libselinux,grub2-common,grub2-tools,grub2-tools-minimal,grubby,file,gettext,gettext-libs,libcroco,libgomp,os-prober,which,diffutils,policycoreutils,libselinux-utils,openssh-clients,libedit,selinux-policy,selinux-policy-targeted,keyutils,libini_config,libverto-libev,nfs-utils,gssproxy,libbasicobjects.libcollection,libev,libevent,libnfsidmap,libpath_utils,libref_array,quota,quota-nls,rpcbind"
 
-# Build a temporary Fedora 28 (x86_64) chroot where we can cross-compile our kernel
 if [ -d "$KERNEL_BUILD_PATH" ]; then
-    echo "Error: Directory $KERNEL_BUILD_PATH exists." >&2
+    echo "Error: Directory $KERNEL_BUILD_PATH exists. Please delete it." >&2
     exit 1
-else
-    mkdir -p "$KERNEL_BUILD_PATH"
 fi
 
+if [ -d "$CHROOT_PATH" ]; then
+    echo "Error: Directory $CHROOT_PATH exists. Please delete it." >&2
+    exit 1
+fi
+
+if [ -d "$INITRAMFS_PATH" ]; then
+    echo "Error: Directory $INITRAMFS_PATH exists. Please delete it." >&2
+    exit 1
+fi
+
+if [ -d "$LIVE_ISO_PATH" ]; then
+    echo "Error: Directory $LIVE_ISO_PATH exists. Please delete it." >&2
+    exit 1
+fi
+
+mkdir -p "$KERNEL_BUILD_PATH"
 dnf -y --use-host-config --releasever=28 --forcearch=x86_64 --disable-repo=* --enable-repo=fedora --installroot=$KERNEL_BUILD_PATH install filesystem
 rm -f $KERNEL_BUILD_PATH/dev/null
 mknod -m 600 $KERNEL_BUILD_PATH/dev/console c 5 1
@@ -46,7 +58,7 @@ cp -f $RESOURCES_PATH/0011-ps3stor-multiple-regions.patch $KERNEL_BUILD_PATH/
 cp -f $RESOURCES_PATH/config-6.8.12-live $KERNEL_BUILD_PATH/linux-6.8.12/.config
 chroot $KERNEL_BUILD_PATH /usr/bin/patch -d /linux-6.8.12 -p1 -i /0011-ps3stor-multiple-regions.patch
 chroot $KERNEL_BUILD_PATH /usr/bin/make ARCH=powerpc CROSS_COMPILE=powerpc64-linux-gnu- -C /linux-6.8.12 olddefconfig
-chroot $KERNEL_BUILD_PATH /usr/bin/make ARCH=powerpc CROSS_COMPILE=powerpc64-linux-gnu- -C /linux-6.8.12 -j1 zImage
+chroot $KERNEL_BUILD_PATH /usr/bin/make ARCH=powerpc CROSS_COMPILE=powerpc64-linux-gnu- -C /linux-6.8.12 -j4 zImage
 chroot $KERNEL_BUILD_PATH /usr/bin/make ARCH=powerpc CROSS_COMPILE=powerpc64-linux-gnu- -C /linux-6.8.12 -j1 modules
 chroot $KERNEL_BUILD_PATH /usr/bin/make ARCH=powerpc CROSS_COMPILE=powerpc64-linux-gnu- -C /linux-6.8.12 modules_install
 rm -f $KERNEL_BUILD_PATH/lib/modules/6.8.12/build
@@ -57,14 +69,7 @@ umount $KERNEL_BUILD_PATH/dev/pts
 umount $KERNEL_BUILD_PATH/dev
 umount $KERNEL_BUILD_PATH/sys
 umount $KERNEL_BUILD_PATH/proc
-
-if [ -d "$CHROOT_PATH" ]; then
-    echo "Error: Directory $CHROOT_PATH exists." >&2
-    exit 1
-else
-    mkdir -p "$CHROOT_PATH"
-fi
-
+mkdir -p "$CHROOT_PATH"
 dnf -y --use-host-config --releasever=28 --forcearch=ppc64 --disable-repo=* --enable-repo=fedora --setopt=install_weak_deps=False --setopt=tsflags=nodocs --installroot=$CHROOT_PATH --exclude=$EXCLUDES install filesystem
 rm -f $CHROOT_PATH/dev/null
 mknod -m 600 $CHROOT_PATH/dev/console c 5 1
@@ -83,13 +88,12 @@ echo "nameserver 8.8.8.8" > $CHROOT_PATH/etc/resolv.conf
 chroot $CHROOT_PATH /usr/bin/dnf --releasever=28 --forcearch=ppc64 clean all
 chroot $CHROOT_PATH /usr/bin/dnf --releasever=28 --forcearch=ppc64 makecache
 chroot $CHROOT_PATH /usr/bin/dnf -y --releasever=28 --forcearch=ppc64 --setopt=install_weak_deps=False --setopt=tsflags=nodocs --exclude=$EXCLUDES groupinstall core
-chroot $CHROOT_PATH /usr/bin/dnf -y --releasever=28 --forcearch=ppc64 --setopt=install_weak_deps=False --setopt=tsflags=nodocs --exclude=$EXCLUDES install udisks2-zram nfs-utils bash-completion wget tar xz wpa_supplicant
+chroot $CHROOT_PATH /usr/bin/dnf -y --releasever=28 --forcearch=ppc64 --setopt=install_weak_deps=False --setopt=tsflags=nodocs --exclude=$EXCLUDES install udisks2-zram bash-completion wget
 chroot $CHROOT_PATH /usr/bin/dnf clean all
 rm -f $CHROOT_PATH/etc/yum.repos.d/*.rpmnew
 mv -f $CHROOT_PATH/etc/nsswitch.conf $CHROOT_PATH/etc/nsswitch.conf.orig
 mv -f $CHROOT_PATH/etc/nsswitch.conf.rpmnew $CHROOT_PATH/etc/nsswitch.conf
 cp -rf $KERNEL_BUILD_PATH/lib/modules/6.8.12 $CHROOT_PATH/lib/modules/
-sed -i 's/SELINUX=enforcing/SELINUX=disabled/g' $CHROOT_PATH/etc/selinux/config
 echo "ps3vram" > $CHROOT_PATH/etc/modules-load.d/ps3vram.conf
 echo 'KERNEL=="ps3vram", ACTION=="add", RUN+="/sbin/mkswap /dev/ps3vram", RUN+="/sbin/swapon -p 200 /dev/ps3vram"' > $CHROOT_PATH/etc/udev/rules.d/10-ps3vram.rules
 chmod 0200 $CHROOT_PATH/etc/shadow
@@ -108,7 +112,7 @@ mkdir $CHROOT_PATH/mnt/target
 cp $RESOURCES_PATH/zram-swap.sh $CHROOT_PATH/usr/sbin/zram-swap.sh
 cp $RESOURCES_PATH/zram-swap.service $CHROOT_PATH/etc/systemd/system/zram-swap.service
 cp $RESOURCES_PATH/ps3linux-install.sh $CHROOT_PATH/usr/sbin/ps3linux-install.sh
-cp $RESOURCES_PATH/ps3linux.repo $CHROOT_PATH/root/ps3linux.repo
+cp $RESOURCES_PATH/ps3linux.repo $CHROOT_PATH/etc/yum.repos.d/ps3linux.repo
 chroot $CHROOT_PATH /usr/bin/systemctl disable auditd.service
 chroot $CHROOT_PATH /usr/bin/systemctl disable dbus-org.fedoraproject.FirewallD1.service
 chroot $CHROOT_PATH /usr/bin/systemctl disable firewalld.service
@@ -124,29 +128,14 @@ umount $CHROOT_PATH/sys
 umount $CHROOT_PATH/proc
 rm -rf $CHROOT_PATH/usr/share/doc
 rm -rf $CHROOT_PATH/usr/share/man
-#rm -rf $CHROOT_PATH/lib/firmware/*
 find $CHROOT_PATH -type f \( -perm -111 -o -name '*.so*' -o -name '*.ko' \) -exec file {} \; | grep 'ELF' | cut -d: -f1 | while read f; do echo "Stripping $f"; powerpc64-linux-gnu-strip --strip-unneeded "$f" || true; done
 find $CHROOT_PATH/usr/lib64 -name '*.a' -delete
-
-if [ -d "$LIVE_ISO_PATH" ]; then
-    echo "Error: Directory $LIVE_ISO_PATH exists." >&2
-    exit 1
-else
-    mkdir -p $LIVE_ISO_PATH/{boot,etc,LiveOS}
-fi
-
-if [ -d "$INITRAMFS_PATH" ]; then
-    echo "Error: Directory $INITRAMFS_PATH exists." >&2
-    exit 1
-else
-    mkdir -p $INITRAMFS_PATH/{dev,lib/modules,mnt/{iso,lower,sysroot,upper},proc,run,sys,tmp,usr/{bin,lib64,sbin}}
-    pushd $INITRAMFS_PATH
-    ln -s usr/bin bin
-    ln -s usr/lib64 lib64
-    ln -s usr/sbin sbin
-    popd
-fi
-
+mkdir -p $INITRAMFS_PATH/{dev,lib/modules,mnt/{iso,lower,sysroot,upper},proc,run,sys,tmp,usr/{bin,lib64,sbin}}
+pushd $INITRAMFS_PATH
+ln -s usr/bin bin
+ln -s usr/lib64 lib64
+ln -s usr/sbin sbin
+popd
 mknod -m 600 $INITRAMFS_PATH/dev/console c 5 1
 mknod -m 666 $INITRAMFS_PATH/dev/null c 1 3
 cp -rf $KERNEL_BUILD_PATH/lib/modules/6.8.12 $INITRAMFS_PATH/lib/modules/
@@ -185,6 +174,7 @@ ln -s libtinfo.so.6.1 libtinfo.so.6
 ln -s ld-2.27.so ld-linux-x86-64.so.2
 popd
 pushd $INITRAMFS_PATH
+mkdir -p $LIVE_ISO_PATH/{boot,etc,LiveOS}
 find . | cpio -H newc -o | gzip > $LIVE_ISO_PATH/boot/initramfs.img
 popd
 cp -f $KERNEL_BUILD_PATH/linux-6.8.12/arch/powerpc/boot/zImage $LIVE_ISO_PATH/boot/vmlinuz
